requirements.txt
fastapi
uvicorn
pydantic
ffmpeg-python